py-script-project/requirements.txt
# Optional: only needed if you want to load YAML config files.
PyYAML

